% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{find_anomaly}
\alias{find_anomaly}
\title{Compute anomalies from the results of prophet}
\usage{
find_anomaly(
  yhat,
  base,
  thresh,
  groups = c("none", "space", "time", "combined"),
  filename_prefix = NULL,
  overwrite = F,
  wopts = NULL
)
}
\arguments{
\item{yhat}{rast. Data cube of yhat predictions from \code{prophet_rast}}

\item{base}{rast. Data cube of "raw" values initially passed into \code{prophet_rast} as \code{r}}

\item{groups}{character. One or more of 'none', 'space', 'time', 'combined' detailing which anamoly data cubes should be returned}

\item{filename_prefix}{file path. Optional. Passed to terra::app, terra::patches, and a possible call to terra::writeRaster. File name (including directory) prefix for output geotiffs}

\item{overwrite}{logical. Optional. Passed to terra::app, terra::patches, and a possible call to terra::writeRaster}

\item{threh}{numeric. Anomly threshold-- Number of standard deviations an estimate must be >= away from the mean}

\item{wotps}{named list. Optional. Passed to terra::app, terra::patches, and a possible call to terra::writeRaster. Options passed to writeRaster}
}
\value{
a list of rast(bricks)
}
\description{
Compute anomalies from the results of prophet
}
