% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic}
\alias{harmonic}
\alias{harmonic_regression}
\title{Compute harmonic regression}
\usage{
harmonic(r, harmonic_value = 2, variable = "all", window = NULL, ...)

harmonic_regression(y, harmonic_value = 2, variable = "all")
}
\arguments{
\item{r}{data cube}

\item{harmonic_value}{numeric. Number of harmonics to consider.}

\item{variable}{character. Name of the variable/result to return. See details for more information.}

\item{window}{numeric. Size of the window of consideration. Chunks the input data cube into cubes with `window` layers/timesteps}

\item{...}{other methods}
}
\description{
Compute harmonic regression
}
\details{
The following values are possibly returned from harmonic:
   0. 'all': return all valid terms
   1. 'A0': Average; base mean change in the system
   2. 'A1': Period specific seasonality
   3. 'A2': Within period seasonality
   4. 'P1': Timing of A1 peak (e.g. peak of season)
   5. 'P2': Timing of A2 peak (e.g. change in dynamics)
}
