% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_grid.R
\name{new_gecm_results}
\alias{new_gecm_results}
\title{Create a new object to store resutls in.
 @param dc data cube from which the reference layer will be created
 @param n_timesteps number of layers in the resulting datacube. usually relates to the time window.
 @param ... unused.
 @importFrom data.table as.data.table ":=" melt .I
 @importFrom terra nlyr values size add}
\usage{
new_gecm_results(dc, n_timesteps = 1, ...)
}
\description{
Create a new object to store resutls in.
 @param dc data cube from which the reference layer will be created
 @param n_timesteps number of layers in the resulting datacube. usually relates to the time window.
 @param ... unused.
 @importFrom data.table as.data.table ":=" melt .I
 @importFrom terra nlyr values size add
}
