% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{output_opts}
\alias{output_opts}
\title{A function to format options for saving rasters while processing}
\usage{
output_opts(fp, name = "rast", overwrite, wopts)
}
\arguments{
\item{fp}{filepath prefix, including folder}

\item{name}{output name}

\item{overwrite}{logical. Optional. Passed to terra::writeRaster (or functions using it)}

\item{wotps}{named list. Optional.  Options passed to writeRaster (or functions using it)}
}
\description{
A function to format options for saving rasters while processing
}
