% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{pixel_fit}
\alias{pixel_fit}
\title{Run prophet and extract predictions}
\usage{
pixel_fit(x, dates, na.thresh = 0.25, outputs, ...)
}
\arguments{
\item{x}{numeric. Vector of values to run prophet on}

\item{dates}{date. Vector of dates same length as \code{x} representing the dates that x correspond to}

\item{na.thresh}{numeric. Minimum proportional of missingness acceptable for a pixel}

\item{outputs}{character. The prophet results to be captured/returned}

\item{...}{other arguments passed to prophet}
}
\description{
Run prophet and extract predictions
}
