% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophet.R
\name{prophet_rast}
\alias{prophet_rast}
\title{Run prophet on a terra data cube}
\usage{
prophet_rast(
  r,
  dates,
  ncores = 1,
  na.thresh = 0.25,
  outputs = c("yhat"),
  ...,
  filename = NULL,
  overwrite = FALSE,
  wopt = list()
)
}
\arguments{
\item{r}{rast. Data cube}

\item{dates}{vector of dates. the date each layer in \code{r} represents}

\item{ncores}{integer. NUmber of processes to use in parallel computing}

\item{na.thresh}{numeric. Minimum proportional of missingness acceptable for a pixel}

\item{outputs}{character. The prophet results to be captured/returned}

\item{...}{other arguments passed to prophet}

\item{filename}{file path. Path to the output file, if using}

\item{overwrite}{logical. Determines whether filename gets overwritten}

\item{wopt}{list. Passed to terra::app}
}
\value{
rast
}
\description{
Run prophet on a terra data cube
}
