% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic}
\alias{harmonic}
\alias{harmonic_regression}
\title{Compute harmonic regression}
\usage{
harmonic(r, harmonic_value = 2, variable = "all", window = NULL, ...)

harmonic_regression(y, harmonic_value = 2, variable = "all")
}
\arguments{
\item{r}{data cube}

\item{harmonic_value}{numeric. Number of harmonics to consider.}

\item{variable}{character. Name of the variable/result to return. See details for more information.}

\item{window}{numeric. Size of the window of consideration. Chunks the input data cube into cubes with `window` layers/timesteps}

\item{...}{other methods}
}
\description{
Compute harmonic regression
}
\details{
The following values are possibly returned from harmonic:
   0. 'all': return all valid terms
   1. 'A': sine term
   2. 'B': cosine term
   3. 'A0': mean/intercept term

Notes to self:
Window option likely assumes equally spaced layers. It doesn't take into account
any dates inherent in the layers.
}
